fix: split server helpers to restore server.py line budget
PR #77 SSE ticket work pushed server.py past the 4320-line CI cap. Extract log redaction, Epic OAuth state, and stream ticket helpers into shared/ modules so main CI passes again.
